+++ clippy_lints/src/renamed_lints.rs
@@ -5,6 +5,7 @@ pub static RENAMED_LINTS: &[(&str, &str)] = &[
     ("clippy::block_in_if_condition_expr", "clippy::blocks_in_if_conditions"),
     ("clippy::block_in_if_condition_stmt", "clippy::blocks_in_if_conditions"),
     ("clippy::box_vec", "clippy::box_collection"),
+    ("clippy::blacklisted_name", "clippy::disallowed_names"),
     ("clippy::const_static_lifetime", "clippy::redundant_static_lifetimes"),
     ("clippy::cyclomatic_complexity", "clippy::cognitive_complexity"),
     ("clippy::disallowed_method", "clippy::disallowed_methods"),